Optimize S3 part upload and checksum computation
Modified uploadPart and checksumPart methods in S3.scala. A chunk of bytes is now converted to an array only once for both uploading to S3 and computing MD5 checksum. This change enhances the efficiency of the code by minimizing the number of times the bytes data is processed, which can yield performance improvements especially for larger files.