add eTagValidation for uploadFileMultipart (#1130)
* add eTagValidation for uploadFileMultipart

* update S3Suite with MultipartETagValidation

* update add test cases for MultipartETagValidation

* compute each chunk md5 digest in parallel and compute the overall digest at the end; same for the max partNumber aka partId

* calc checksum of chunk in one shot

* wrap the part processing state into case class to avoid _.1 things

* merged upstream

* fixed bad merge of upstream

* additional fixed bad merge of upstream

* separate out unit test for MultipartETagValidation from S3 integration test

* Optimized code and added benchmarks for S3 multi-part uploads

Refactored `S3.scala` to optimize `checksumPart` and `formatChecksum` functions, thereby improving code readability and performance. Added `S3Benchmark.scala` and `S3OpsStub.scala` for benchmarking S3 multi-part uploads.

* have MultipartETagValidation return a value.
This way it can be verified and eventually incorporated in the return from uploadFileMultipart in future versions.

* Optimize S3 part upload and checksum computation

Modified uploadPart and checksumPart methods in S3.scala. A chunk of bytes is now converted to an array only once for both uploading to S3 and computing MD5 checksum. This change enhances the efficiency of the code by minimizing the number of times the bytes data is processed, which can yield performance improvements especially for larger files.

* Refactor PartProcessingOutcome to use optional digest

Changed the 'digest' field in the PartProcessingOutcome case class to be optional due to the multipartETagValidation changes. Modified the getOverallChecksum and upload parts methods accordingly, to handle Option[PartDigest] safely and efficiently. This is done to prevent unnecessary digest calculations during multipart upload when checksum validation is not necessary

* Refactor PartProcessingOutcome to use optional digest

Changed the 'digest' field in the PartProcessingOutcome case class to be optional due to the multipartETagValidation changes. Modified the getOverallChecksum and upload parts methods accordingly, to handle Option[PartDigest] safely and efficiently. This is done to prevent unnecessary digest calculations during multipart upload when checksum validation is not necessary

---------

Co-authored-by: Shawn Garner <garnershawn@johndeere.com>
Co-authored-by: dmytro.obodowsky <dmytro.obodowsky@disney.com>